+++ src/components/Layout/Footer/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { ReactComponent as HeartIcon } from '../../../assets/img/heart.svg';
 
-const Footer = () => {
+export const Footer = () => {
   return (
     <footer className='text-center mt-2 mb-3'>
       <div>
@@ -18,5 +18,3 @@ const Footer = () => {
     </footer>
   );
 };
-
-export default Footer;
