+++ src/components/Layout/Navbar/index.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { useGetAccountInfo } from '@elrondnetwork/dapp-core/hooks';
+import { useGetIsLoggedIn } from '@elrondnetwork/dapp-core/hooks';
 import { logout } from '@elrondnetwork/dapp-core/utils';
 import { Navbar as BsNavbar, NavItem, Nav } from 'react-bootstrap';
 import { Link } from 'react-router-dom';
@@ -7,14 +7,12 @@ import { dAppName } from 'config';
 import { routeNames } from 'routes';
 import { ReactComponent as ElrondLogo } from './../../../assets/img/elrond.svg';
 
-const Navbar = () => {
-  const { address } = useGetAccountInfo();
+export const Navbar = () => {
+  const isLoggedIn = useGetIsLoggedIn();
 
   const handleLogout = () => {
     logout(`${window.location.origin}/unlock`);
   };
-
-  const isLoggedIn = Boolean(address);
 
   return (
     <BsNavbar className='bg-white border-bottom px-4 py-3'>
@@ -40,5 +38,3 @@ const Navbar = () => {
     </BsNavbar>
   );
 };
-
-export default Navbar;
