+++ src/components/Layout/index.tsx
@@ -1,26 +1,1 @@
-import React from 'react';
-import { AuthenticatedRoutesWrapper } from '@elrondnetwork/dapp-core/wrappers';
-import { useLocation } from 'react-router-dom';
-import routes, { routeNames } from 'routes';
-import Footer from './Footer';
-import Navbar from './Navbar';
-
-const Layout = ({ children }: { children: React.ReactNode }) => {
-  const { search } = useLocation();
-  return (
-    <div className='bg-light d-flex flex-column flex-fill wrapper'>
-      <Navbar />
-      <main className='d-flex flex-column flex-grow-1'>
-        <AuthenticatedRoutesWrapper
-          routes={routes}
-          unlockRoute={`${routeNames.unlock}${search}`}
-        >
-          {children}
-        </AuthenticatedRoutesWrapper>
-      </main>
-      <Footer />
-    </div>
-  );
-};
-
-export default Layout;
+export * from './Layout';
