+++ src/components/index.tsx
@@ -1,0 +1,3 @@
+export * from './Layout';
+export * from './PageTitle';
+export * from './AuthRedirectWrapper';
