+++ src/index.tsx
@@ -2,7 +2,7 @@ import './index.css';
 import './assets/sass/theme.scss';
 import React from 'react';
 import { createRoot } from 'react-dom/client';
-import App from './App';
+import { App } from './App';
 
 const container = document.getElementById('root');
 const root = createRoot(container as HTMLElement);
