+++ src/pages/Dashboard/Actions/helpers/index.ts
@@ -1,0 +1,2 @@
+export * from './useGetPingAmount';
+export * from './useGetTimeToPong';
