+++ src/pages/Dashboard/Transactions/txStatus.tsx
@@ -1,8 +1,6 @@
-const txStatus = {
+export const txStatus = {
   pending: 'Pending',
   notExecuted: 'Not Executed',
   success: 'Success',
   fail: 'Failed'
 };
-
-export default txStatus;
