--- src/pages/Dashboard/helpers/index.ts
@@ -1,1 +1,0 @@
-export * from './getSmartContract';
