+++ src/pages/Home.tsx
@@ -3,7 +3,7 @@ import { Link } from 'react-router-dom';
 import { dAppName } from 'config';
 import { routeNames } from 'routes';
 
-const Home = () => {
+export const Home = () => {
   return (
     <div className='d-flex flex-fill align-items-center container'>
       <div className='row w-100'>
@@ -33,5 +33,3 @@ const Home = () => {
     </div>
   );
 };
-
-export default Home;
