+++ src/pages/UnlockPage.tsx
@@ -8,7 +8,7 @@ import {
 import { walletConnectV2ProjectId } from 'config';
 import { routeNames } from 'routes';
 
-export const UnlockRoute: () => JSX.Element = () => {
+export const UnlockPage = () => {
   return (
     <div className='home d-flex flex-fill align-items-center'>
       <div className='m-auto' data-testid='unlockPage'>
@@ -45,5 +45,3 @@ export const UnlockRoute: () => JSX.Element = () => {
     </div>
   );
 };
-
-export default UnlockRoute;
