+++ src/pages/index.tsx
@@ -1,0 +1,5 @@
+export * from './Dashboard';
+export * from './Home';
+export * from './PageNotFound';
+export * from './UnlockPage';
+export * from './Transaction';
